Split option parsing
