+++ cli/run/loadConfigFile.ts
@@ -4,8 +4,8 @@ import { pathToFileURL } from 'url';
 import * as rollup from '../../src/node-entry';
 import { MergedRollupOptions } from '../../src/rollup/types';
 import { error } from '../../src/utils/error';
-import { mergeOptions } from '../../src/utils/mergeOptions';
-import { GenericConfigObject } from '../../src/utils/parseOptions';
+import { mergeOptions } from '../../src/utils/options/mergeOptions';
+import { GenericConfigObject } from '../../src/utils/options/parseOptions';
 import relativeId from '../../src/utils/relativeId';
 import { stderr } from '../logging';
 import batchWarnings, { BatchWarnings } from './batchWarnings';
